Hide `descriptortype` from users with a new `UPB_PRIVATE()` macro
We are running into more and more situations where users are accessing private struct members directly.  This is difficult to prevent or detect: the only hint to a user that these struct members are private is that they are defined in a file like `msg_internal.h`.  But this is easy to miss, particularly if the IDE will happily auto-complete members as you type.

The `UPB_PRIVATE()` macro makes these names unmistakably private.  The actual name will be something like `descriptortype__do_not_use__i_give_permission_to_break_me`.  The macro is defined in a header that users cannot `#include`, and our macro state will not leak to the user thanks to `port_undef.inc`.

If this experiment goes well, we should hide all members of all internal-only structs with `UPB_PRIVATE()`.

Note that this only applies to structs that:

1. Are defined in headers (structs defined in `.c` files are already hidden)
2. Are transitively included in public headers (structs in fully-internal headers are already hidden).

This would primarily apply to MiniTables and their members.

PiperOrigin-RevId: 477306585