mini table field_parser is now an untyped pointer.
The field_parser function pointer was requiring knowledge of the wire decoder, whic is a higher-level object. The mini table itself doesn't care what kind of function
this is, it's merely providing storage. So, make it a (void*) and just cast it
upon invocation as appropriate.

Tested both with and without UPB_ENABLE_FASTTABLE set.

PiperOrigin-RevId: 488085173